Tarde 25 abril

+++ buscarMisPokemonPorNombre.c
@@ -3,11 +3,12 @@
 #include <string.h>
 #include <stdbool.h>
 #include "Map.h"
+
 typedef struct{
 	int id;
 	char* nombre;
-	int PC;
-	int PS;
+	float PC;
+	float PS;
 	char* sexo;
 }Pokemon_usuario;
